// packages/external-api/src/holidays/generated/models/HolidayDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* eslint-disable */
export type HolidayDto = {
  end?: string;
  name?: string;
  slug?: string;
  start?: string;
  stateCode?: HolidayDto.stateCode;
  year?: number;
};
export namespace HolidayDto {
  export enum stateCode {
    BB = "BB",
    BE = "BE",
    BW = "BW",
    BY = "BY",
    HB = "HB",
    HE = "HE",
    HH = "HH",
    MV = "MV",
    NI = "NI",
    NW = "NW",
    RP = "RP",
    SH = "SH",
    SL = "SL",
    SN = "SN",
    ST = "ST",
    TH = "TH",
  }
}